feat: update reviews, comments views

+++ products/views.py
@@ -11,7 +11,40 @@ from products.models  import Product, ProductImage, Category, CategoryImage, Com
 from users.models     import User
 from users.decorator  import log_in_decorator
 
-class ProductView(View):
+
+class ProductListView(View):
+    def get(self, request):
+        try:
+            category_id = request.GET.get('categoryId', None)
+            searching   = request.GET.get('productName', None)
+            offset      = request.GET.get('offset', 0)
+            limit       = request.GET.get('limit', 30)
+
+            filter_condition = Q()
+
+            if category_id:
+                filter_condition &= Q(category_id=category_id)
+
+            if searching:
+                filter_condition &= Q(name__icontains=searching)
+
+            products = Product.objects.filter(filter_condition).order_by('?')[offset:offset+limit]
+
+            product_list = [{
+                'category_id'     : product.category.id,
+                'product_id'      : product.id,
+                'name'            : product.name,
+                'price'           : product.price,
+                'description_tag' : product.description_tag,
+                'products_image'  : product.productimage_set.first().image_url,
+            } for product in products]
+
+            return JsonResponse({'product_list' : product_list}, status = 200)
+
+        except ValueError:
+            return JsonResponse({'message':'VALUE_ERROR'}, status = 400)
+
+class ProductDetailView(View):
     def get(self, request, product_id):
         try:
             products = Product.objects.filter(id = product_id)
@@ -59,7 +92,6 @@ class CommentView(View):
             data         = json.loads(request.body)
             user         = request.user
             content      = data.get('content', None)
-            product_id   = data.get('product_id', None)
 
             if not Product.objects.filter(id=product_id).exists():
                 return JsonResponse({'message': 'PRODUCT_DOES_NOT_EXIST'}, status=404)
@@ -80,7 +112,7 @@ class CommentView(View):
             return JsonResponse({'message':'COMMENT_DOES_NOT_EXIST'}, status=404)
 
         comment_list = [{
-            "username"  : User.objects.get(id=comment.user.id).username,
+            "name"  : User.objects.get(id=comment.user.id).name,
             "content"   : comment.content,
             "create_at" : comment.created_at
             } for comment in Comment.objects.filter(product_id=product_id)
@@ -89,16 +121,14 @@ class CommentView(View):
         return JsonResponse({'data':comment_list}, status=200)
 
     @log_in_decorator
-    def delete(self, request, comment_id):
-        user = request.user
-
-        if Comment.objects.get(id=comment_id).exists():
+    def delete(self, request, product_id, comment_id):
+        user_id    = request.user.id
+        product_id = product_id
+        
+        if not Comment.objects.filter(id=comment_id, product_id=product_id, user_id=user_id).exists():
             return JsonResponse({'message':'COMMENT_DOES_NOT_EXIST'}, status=404)
 
-        if user != Comment.user:
-            return JsonResponse({'message':'INVALID_USER'}, status=401)
-
-        Comment.objects.filter(id=comment_id).delete()
+        Comment.objects.filter(id=comment_id, product_id=product_id, user_id=user_id).first().delete()
         return JsonResponse({'message': 'SUCCESS'}, status=204)
 
 class SubscribeView(View):
@@ -132,34 +162,3 @@ class SubscribeView(View):
             },
             status = 200)
 
-class ProductListView(View):
-    def get(self, request):
-        try:
-            category_id = request.GET.get('categoryId', None)
-            searching   = request.GET.get('productName', None)
-            offset      = request.GET.get('offset', 0)
-            limit       = request.GET.get('limit', 30)
-
-            filter_condition = Q()
-
-            if category_id:
-                filter_condition &= Q(category_id=category_id)
-
-            if searching:
-                filter_condition &= Q(name__icontains=searching)
-
-            products = Product.objects.filter(filter_condition).order_by('?')[offset:offset+limit]
-
-            product_list = [{
-                'category_id'     : product.category.id,
-                'product_id'      : product.id,
-                'name'            : product.name,
-                'price'           : product.price,
-                'description_tag' : product.description_tag,
-                'products_image'  : product.productimage_set.first().image_url,
-            } for product in products]
-
-            return JsonResponse({'product_list' : product_list}, status = 200)
-
-        except ValueError:
-            return JsonResponse({'message':'VALUE_ERROR'}, status = 400)
